k8s/k8s/creator/src/requirements.txt
google-cloud-iam
kopf
google-cloud
logging
google-api-python-client
